+++ src/symbolsinformation/methodInformation.ts
@@ -1,4 +1,5 @@
 export class MethodInformation {
     name: string | undefined;
     header: string | undefined;
+    accessModifier: string | undefined;
 }